Rollback inadvertent commit to filter_test.c
git-svn-id: http://svn.research.cc.gatech.edu/kaos/evpath/trunk@20326 8ba6a8b9-845b-0410-94f0-974be2f5df47

+++ tests/filter_test.c
@@ -84,7 +84,7 @@ static FMStructDescRec simple_format_list[] =
 static FMField filter_field_list[] =
 {
     {"integer_field", "integer",
-     sizeof(short), FMOffset(simple_rec_ptr, integer_field)},
+     sizeof(int), FMOffset(simple_rec_ptr, integer_field)},
     {NULL, NULL, 0, 0}
 };
